backend/requirements.txt
flask
flask-corss